refactor CoreTests directories and file names
