+++ src/backend/CoreTests/Actions/GetProducts/GetProductsFixture.cs
@@ -1,12 +1,11 @@
 ﻿using Core.Database.Enums;
 using Core.Database.Models.Auth;
-using Core.Database.Models.Core;
 using Core.Database.Seeders;
 using Core.Setup.Auth0;
 
-namespace CoreTests.Actions.MyNotificationsGetListLast30DaysTest
+namespace CoreTests.Actions.GetProducts
 {
-    public static class MyNotificationsGetListLast30DaysFixture
+    public static class GetProductsFixture
     {
         public static User CreateUser(string firstName, string lastName) => new()
         {
@@ -29,14 +28,26 @@ namespace CoreTests.Actions.MyNotificationsGetListLast30DaysTest
             UserId = user.Id
         };
 
-        public static Notification CreateNotifaction(User user, long daysPastFromToday, string partMessage) => new()
+        public static UserRole CreateUserRole(User user, long roleId) => new()
         {
-            UserId = user.Id,
+            RoleId = roleId,
             User = user,
-            CreationDate = DateTime.UtcNow.AddDays(-daysPastFromToday),
-            Message = "You've reached level 10 donor! " + partMessage,
             UpdateUserId = UserSeeder.ServiceUser.Id,
-            UpdatedAt = DateTime.UtcNow
+            UpdatedAt = DateTime.UtcNow,
+            Status = DbEntityStatus.Active
+        };
+
+        public static CurrentUserInfo GetCurrentUserInfo(Auth0User auth0User) => new()
+        {
+            Email = auth0User.Email,
+            EmailVerified = true,
+            Identifier = auth0User.Identifier,
+            GivenName = auth0User.FirstName,
+            FamilyName = auth0User.LastName,
+            Locale = "pl",
+            Name = auth0User.FirstName + auth0User.LastName,
+            Nickname = auth0User.Nickname,
+            UpdatedAt = DateTime.UtcNow,
         };
     }
 }
