+++ src/backend/CoreTests/Services/Counter/CounterServiceTest.cs
@@ -7,10 +7,10 @@ using Core.Setup.Enums;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using TestsBase;
 
-namespace CoreTests.Services
+namespace CoreTests.Services.Counter
 {
     [TestClass()]
-    public class CounterServiceTests
+    public class CounterServiceTest
     {
         private readonly List<Module> _usedModules = new()
         {
@@ -18,7 +18,7 @@ namespace CoreTests.Services
         };
 
         [TestMethod()]
-        public async Task CounterServiceTests_DonationTest_ReturnCounterOfDonationTable()
+        public async Task CounterServiceTest_DonationTest_ReturnCounterOfDonationTable()
         {
             using var toolkit = new TestsToolkit(_usedModules);
             var context = toolkit.Resolve<AppDbContext>();
@@ -35,7 +35,7 @@ namespace CoreTests.Services
         }
 
         [TestMethod()]
-        public async Task CounterServiceTests_UpdateDonationCounter_ReturnUpdatedCounterOfDonationTable()
+        public async Task CounterServiceTest_UpdateDonationCounter_ReturnUpdatedCounterOfDonationTable()
         {
             using var toolkit = new TestsToolkit(_usedModules);
             var context = toolkit.Resolve<AppDbContext>();
@@ -52,7 +52,7 @@ namespace CoreTests.Services
         }
 
         [TestMethod()]
-        public async Task CounterServiceTests_OrderTest_ReturnCounterOfOrderTable()
+        public async Task CounterServiceTest_OrderTest_ReturnCounterOfOrderTable()
         {
             using var toolkit = new TestsToolkit(_usedModules);
             var context = toolkit.Resolve<AppDbContext>();
@@ -69,7 +69,7 @@ namespace CoreTests.Services
         }
 
         [TestMethod()]
-        public async Task CounterServiceTests_UpdateOrderCounter_ReturnUpdatedCounterOfOrderTable()
+        public async Task CounterServiceTest_UpdateOrderCounter_ReturnUpdatedCounterOfOrderTable()
         {
             using var toolkit = new TestsToolkit(_usedModules);
             var context = toolkit.Resolve<AppDbContext>();
@@ -86,7 +86,7 @@ namespace CoreTests.Services
         }
 
         [TestMethod()]
-        public async Task CounterServiceTests_DeliveryTest_ReturnCounterOfDeliveryTable()
+        public async Task CounterServiceTest_DeliveryTest_ReturnCounterOfDeliveryTable()
         {
             using var toolkit = new TestsToolkit(_usedModules);
             var context = toolkit.Resolve<AppDbContext>();
@@ -103,7 +103,7 @@ namespace CoreTests.Services
         }
 
         [TestMethod()]
-        public async Task CounterServiceTests_UpdateDeliveryCounter_ReturnUpdatedCounterOfDeliveryTable()
+        public async Task CounterServiceTest_UpdateDeliveryCounter_ReturnUpdatedCounterOfDeliveryTable()
         {
             using var toolkit = new TestsToolkit(_usedModules);
             var context = toolkit.Resolve<AppDbContext>();
